Fix swagger parse problem

+++ Api/TodoListApi/Controllers/TasksController.cs
@@ -10,10 +10,29 @@ namespace TodoListApi.Controllers
     {
 
         [HttpGet()]
-        public async Task<IActionResult> Get()
+        public async Task<IActionResult> Get([FromQuery] TasksSearchRequestModel model)
         {
-            var response = await taskService.SearchTasksAsync(new TasksSearchRequestModel());
+            var response = await taskService.SearchTasksAsync(model);
             return Ok(response);
+        }
+
+        [HttpGet("{id}")]
+        public async Task<IActionResult> GetById([FromRoute(Name = "id")] string stringId)
+        {
+            if (!long.TryParse(stringId, out long id))
+            {
+                return BadRequest(new { message = "Invalid ID format" });
+            }
+
+            try
+            {
+                var task = await taskService.GetTaskByIdAsync(id);
+                return Ok(task);
+            }
+            catch (KeyNotFoundException ex)
+            {
+                return NotFound(new { message = ex.Message });
+            }
         }
 
         [HttpPost]
@@ -22,7 +41,7 @@ namespace TodoListApi.Controllers
             try
             {
                 var result = await taskService.CreateTaskAsync(model);
-                return CreatedAtAction(nameof(Get), new { id = result.Id }, result);
+                return CreatedAtAction(nameof(GetById), new { id = result.Id }, result);
             }
             catch (ArgumentException ex)
             {
@@ -31,8 +50,13 @@ namespace TodoListApi.Controllers
         }
 
         [HttpDelete("{id}")]
-        public async Task<IActionResult> Delete(long id)
+        public async Task<IActionResult> Delete([FromRoute(Name = "id")] string stringId)
         {
+            if (!long.TryParse(stringId, out long id))
+            {
+                return BadRequest(new { message = "Invalid ID format" });
+            }
+
             try
             {
                 await taskService.DeleteTaskAsync(id);
